Make DCO allow CC0 contributions
The Developer Certificate of Origin requires pull requests to be
submitted under an "open source license." At least one of Creative
Commons's repos is available under CC0 [1]. CC0 is a public domain
dedication, not a license [2]. Even if it was, it's not clear whether or
not it would be an open source one since CC0 tries to waive copyright
[3][4].

The DCO itself doesn't allow modifications, so I included a note above
it. The text of the note was based on a suggestion from @TimidRobot [5].
I made a few changes compared to that original suggestion:
  i. I changed the sentence so it wasn't about CC's interpretation of
     the language. The DCO is supposed to be a statement made by
     contributors, not a statement made by the one receiving the
     contributions.
  ii. There's a few phrases that are difficult to fix in part (b). They
      reference "that license" and "a different license." I added an
      extra part to make sure that the phrase "that license" can refer
      to a public domain dedication.
  iii. I removed the "for use with open content" part since everything
       in the public domain is open content [6].
  iv. I tried to make the sentence's structure more parallel.

References:
  1. <https://github.com/creativecommons/cc-licenses-data#copying>
  2. <https://creativecommons.org/faq/#how-do-cc-licenses-operate>
  3. <https://opensource.org/faq#public-domain>
  4. <https://opensource.org/faq#cc-zero>
  5. <https://github.com/creativecommons/.github/issues/24#issuecomment-970998536>
  6. <https://opencontent.org/definition/>

Fixes #24.